+++ src/Psr6/Taggable/InvalidArgumentException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MatthiasMullie\Scrapbook\Psr6\Taggable;
+
+class InvalidArgumentException extends \MatthiasMullie\Scrapbook\Psr6\InvalidArgumentException
+{
+}